+++ CHANGELOG.md
@@ -1,2 +1,16 @@
 ## [0.4.0] 
 * 修改readme.
+
+## [0.5.0] 
+* 添加reloadData方法.
+
+## [0.6.0] 
+* 修改beginRefresh方法显示头部开始转圈圈.
+
+## [0.7.0] 
+* 修复在没有mounted情况下setState的bug.
+* 修复item没有铺满全屏时,listView响应范围过小的bug.
+
+## [0.8.0]
+* 修复'data.success'为null情况下的直接判断的bug.
+* 添加refreshIndicatorColor和refreshIndicatorBackgroundColor属性